requirements.txt
asgiref==3.8.1
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
GDAL==3.9.2
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.1
gunicorn==20.1.0
whitenoise==6.4.0
django-cors-headers==4.0.0
djangorestframework-gis==1.1
numpy==2.1